Cut long chains in task debug information
When loading long chains of debug data, the browser load is increasing
drastically. This is either due to an excessive amount of prints or
to the binary value of uploaded files that are printed out in the debug
dialog. This commit cuts all the chains that exceeds 1000 characters
(if this is a string) or bytes (if this is a byte chain).